refactor(webapp): extract meet now submit orchestration
Move the meet now submit flow into a plain submitMeetNow function with
explicit dependencies, following the same pattern as useScheduleMeetingSubmit.
The hook now only wires React state and dependencies.
